Move Layout to its own file, to help shrink __init__.py
For now, we import it directly where it was defined before, so that this can be done piecewise
